# fire-backend/contracts/openapi.json
{
  "info": {
    "title": "Smart Fire Evacuation API",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/sos": {
      "post": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "tags": [
          "incidents"
        ],
        "summary": "Need help / trapped. Uses the device's latest position",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SosIn"
              }
            }
          }
        }
      }
    },
...
    "/api/me/status": {
      "post": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "tags": [
          "incidents"
        ],
        "summary": "Resident reports evacuating / safe / needs help",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MyStatusIn"
              }
            }
          }
        }
      }
    },
...
    "/api/sos/{sos_id}/resolve": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "sos_id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "tags": [
          "incidents"
        ],
        "summary": "Resolve Sos"
      }
    },
...
    "/api/incidents/{incident_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "tags": [
          "incidents"
        ],
        "summary": "Get Incident"
      }
    },
...
    "/api/buildings/{building_id}/sos": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "building_id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "tags": [
          "incidents"
        ],
        "summary": "List Sos"
      }
    },
...
    "/api/incidents/{incident_id}/status": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "tags": [
          "incidents"
        ],
        "summary": "Declare all-clear or close an incident (computes post-incident metrics)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentStatusIn"
              }
            }
          }
        }
      }
    },
    "/api/incidents/{incident_id}/events": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "tags": [
          "incidents"
        ],
        "summary": "Event log for replay. Filters: ?types=snapshot,route.assigned&after_id=0&limit=2000"
      }
    },
...
    "/api/buildings/{building_id}/incidents": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "building_id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "tags": [
          "incidents"
        ],
        "summary": "List Incidents"
      },
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "building_id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "tags": [
          "incidents"
        ],
        "summary": "Start a fire incident or a drill manually; residents are alerted",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentIn"
              }
            }
          }
        }
      }
    },
...
    "/api/incidents/{incident_id}/export.csv": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "tags": [
          "incidents"
        ],
        "summary": "Export Csv"
      }
    },
...
  },
  "openapi": "3.1.0",
  "components": {
    "schemas": {
      "ValidationError": {
        "properties": {
          "detail": {
            "type": "object",
            "properties": {
              "<location>": {
                "type": "object",
                "properties": {
                  "<field_name>": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "SosIn": {
        "properties": {
          "kind": {
            "default": "help",
            "pattern": "^(help|trapped|medical)$",
            "title": "Kind",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Note"
          }
        },
        "title": "SosIn",
        "type": "object"
      },
...
      "MyStatusIn.ParticipantStatus": {
        "enum": [
          "unknown",
          "evacuating",
          "safe",
          "needs_help"
        ],
        "title": "ParticipantStatus",
        "type": "string"
      },
      "MyStatusIn": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/MyStatusIn.ParticipantStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "MyStatusIn",
        "type": "object"
      },
...
      "HTTPError": {
        "properties": {
          "detail": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "IncidentStatusIn.IncidentStatus": {
        "enum": [
          "detected",
          "evacuating",
          "all_clear",
          "closed"
        ],
        "title": "IncidentStatus",
        "type": "string"
      },
      "IncidentStatusIn": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/IncidentStatusIn.IncidentStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "IncidentStatusIn",
        "type": "object"
      },
...
      "IncidentIn": {
        "properties": {
          "kind": {
            "default": "drill",
            "pattern": "^(fire|drill)$",
            "title": "Kind",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Note"
          }
        },
        "title": "IncidentIn",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}